openvpn: Updated to 2.7.1.

+++ release/src/router/openvpn/src/openvpn/gremlin.c
@@ -5,7 +5,7 @@
  *             packet encryption, packet authentication, and
  *             packet compression.
  *
- *  Copyright (C) 2002-2024 OpenVPN Inc <sales@openvpn.net>
+ *  Copyright (C) 2002-2026 OpenVPN Inc <sales@openvpn.net>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License version 2
@@ -17,8 +17,7 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License along
- *  with this program; if not, write to the Free Software Foundation, Inc.,
- *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *  with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
 /*
@@ -65,28 +64,29 @@ static const int corrupt_freq[] = { 500, 100, 50 };
  * When network goes up, it will be up for between
  * UP_LOW and UP_HIGH seconds.
  */
-static const int up_low[] =  {  60, 10,  5 };
+static const int up_low[] = { 60, 10, 5 };
 static const int up_high[] = { 600, 60, 10 };
 
 /*
  * When network goes down, it will be down for between
  * DOWN_LOW and DOWN_HIGH seconds.
  */
-static const int down_low[] =  {  5, 10,  10 };
+static const int down_low[] = { 5, 10, 10 };
 static const int down_high[] = { 10, 60, 120 };
 
 /*
  * Packet flood levels:
  *  { number of packets, packet size }
  */
-static const struct packet_flood_parms packet_flood_data[] =
-{{10, 100}, {10, 1500}, {100, 1500}};
+static const struct packet_flood_parms packet_flood_data[] = { { 10, 100 },
+                                                               { 10, 1500 },
+                                                               { 100, 1500 } };
 
 struct packet_flood_parms
 get_packet_flood_parms(int level)
 {
     ASSERT(level > 0 && level < 4);
-    return packet_flood_data [level - 1];
+    return packet_flood_data[level - 1];
 }
 
 /*
@@ -107,7 +107,7 @@ roll(int low, int high)
 {
     int ret;
     ASSERT(low <= high);
-    ret = low + (get_random() % (high - low + 1));
+    ret = low + (int)(get_random() % (high - low + 1));
     ASSERT(ret >= low && ret <= high);
     return ret;
 }
@@ -148,18 +148,16 @@ ask_gremlin(int flags)
             int delta;
             if (up)
             {
-                delta = roll(down_low[up_down_level-1], down_high[up_down_level-1]);
+                delta = roll(down_low[up_down_level - 1], down_high[up_down_level - 1]);
                 up = false;
             }
             else
             {
-                delta = roll(up_low[up_down_level-1], up_high[up_down_level-1]);
+                delta = roll(up_low[up_down_level - 1], up_high[up_down_level - 1]);
                 up = true;
             }
 
-            msg(D_GREMLIN,
-                "GREMLIN: CONNECTION GOING %s FOR %d SECONDS",
-                (up ? "UP" : "DOWN"),
+            msg(D_GREMLIN, "GREMLIN: CONNECTION GOING %s FOR %d SECONDS", (up ? "UP" : "DOWN"),
                 delta);
             next = now + delta;
         }
@@ -167,7 +165,7 @@ ask_gremlin(int flags)
 
     if (drop_level)
     {
-        if (up && flip(drop_freq[drop_level-1]))
+        if (up && flip(drop_freq[drop_level - 1]))
         {
             dmsg(D_GREMLIN_VERBOSE, "GREMLIN: Random packet drop");
             return false;
@@ -186,13 +184,13 @@ corrupt_gremlin(struct buffer *buf, int flags)
     const int corrupt_level = GREMLIN_CORRUPT_LEVEL(flags);
     if (corrupt_level)
     {
-        if (flip(corrupt_freq[corrupt_level-1]))
+        if (flip(corrupt_freq[corrupt_level - 1]))
         {
             do
             {
                 if (buf->len > 0)
                 {
-                    uint8_t r = roll(0, 255);
+                    uint8_t r = (uint8_t)roll(0, 255);
                     int method = roll(0, 5);
 
                     switch (method)
@@ -227,8 +225,9 @@ corrupt_gremlin(struct buffer *buf, int flags)
                 {
                     break;
                 }
-            } while (flip(2));  /* a 50% chance we will corrupt again */
+            } while (flip(2)); /* a 50% chance we will corrupt again */
         }
     }
 }
+
 #endif /* ifdef ENABLE_DEBUG */
